Refactored Crud/QueryParams
