+++ src/components/QueryParams/index.vue
@@ -39,4 +39,107 @@ export default {
     //parse filters for upload?
 
 }
+
+/*
+CreateGetRouteParams takes the objects generated by the modal windows, parses them and preprocess the data type. If all values
+are valid its included in the final object. Else its excluded from the final object sent to the server.
+*/
+function CreateGetRouteParams(queryParams) {
+    /*
+    Pulls data rules and intializes paramters. queryParams is an object that contains the current values
+    from all the modal windows.
+    req_body from client has the following structure
+    req.body['rules'] = [{}] //rules contain informations such as include read only.
+    req.body['where'] = [{}]
+    req.body['order_by'] = [{}]
+    req.body['pagination'] = {}
+    */
+    var req_body = {'where': [], 'order_by': [], 'pagination': {}, 'rules': [{}] }
+    var where = queryParams['where']
+    var order_by = queryParams['order_by']
+    var pagination = queryParams['pagination']
+    ParseWhereObject(req_body, where)
+    ParseOrderBy(req_body, order_by)
+    ParsePagination(req_body, pagination)
+    return req_body
+}
+
+function ParsePagination(req_body, pagination) {
+    var lx = parseInt(pagination['limit'])
+    var offset = parseInt(pagination['offset'])
+    if (isNaN(lx) ) {lx = 1000} else if (lx < 10 ) {lx = 10} else if (lx > 5000) {lx = 5000}
+    if (isNaN(offset) ) {offset = 0} else if (offset < 0 ) {offset = 0}
+    var pgx = req_body['pagination']
+    pgx['limit']  = lx
+    pgx['offset'] = offset
+}
+
+function ParseOrderBy(req_body, order_by) {
+    //if column name is empty or sort_by is empty continue
+    var order_by_list = req_body['order_by']
+    for (var i =0; i < order_by.length; i++) {
+        var ox = order_by[i]
+        var cn = ox['variable_name'].trim()
+        var sb = ox['sort_order'].toUpperCase()
+        if (['ASC', 'DESC'].includes(sb) && cn !== "" ) {
+            order_by_list.push({'variable_name': cn, 'sort_order': sb})
+        }
+    }
+}
+
+
+function ParseWhereObject(req_body, where_list) {
+    /*
+    where_list: contains the where statemenents generated by the modal windows where, order_by and pagination
+    where_list_out: container object for req_body. contains processed where statements to send to the server
+    */
+    var where_list_out = req_body['where']
+    for(var i =0; i< where_list.length; i++) {
+        var where_statement = where_list[i]
+        var filter_active = where_statement['filter_active']
+        if (!filter_active) {continue }
+        var dt = where_statement['data_type']
+        if (dt === 'permissions') {
+            QueryPermissionsParse (where_list_out, where_statement)
+        } else if (dt === 'quick_filter') {
+            QueryQuickFilterParse (where_list_out, where_statement)
+        } else if (dt === 'integer') {
+            QueryNumberParse (where_list_out, where_statement, dt)
+        } else if (dt === 'float') {
+            QueryNumberParse (where_list_out, where_statement, dt)
+        } else if (dt === 'string') {
+            QueryStringParse (where_list_out, where_statement)
+        } else if (dt === 'boolean') {
+            QueryBooleanParse (where_list_out, where_statement, dt)
+        } else if (dt === 'date') {
+            QueryDateParse (where_list_out, where_statement)
+        }
+    }
+}
+
+
+//Create Initial Paramters for pagination
+function PaginationInitialization(pagination_rules) {
+    var pagination = {}
+    if (pagination_rules.hasOwnProperty('offset')) {
+        pagination['offset'] = parseInt(pagination_rules['offset'])
+        if (isNaN(pagination['offset'])) {
+            pagination['offset'] = 0
+        } else { if (pagination['offset'] < 0) { pagination['offset'] = 0 } }
+
+    } else { pagination['offset'] = 0 }
+
+    if (pagination_rules.hasOwnProperty('limit')) {
+        pagination['limit'] = parseInt(pagination_rules['limit'])
+        if (isNaN(pagination['limit'])) {
+            pagination['limit'] = 1000
+        } else { 
+            if (pagination['limit'] < 10) { pagination['limit'] = 10 } 
+            else if ( pagination['limit'] > 5000 ) { pagination['limit'] = 5000 } 
+        }
+    } else { pagination['limit'] = 1000 }
+
+    return pagination
+}
+
 </script>
